refactor: url viewer page
